Update VideoRenderer component
