+++ src/atoms/channels.ts
@@ -1,4 +1,6 @@
 import { atom } from 'jotai';
-import { Channel } from '../models/Channel';
+import { Channel, ChannelSelection } from '../models/Channel';
 
 export const channelsAtom = atom([] as Channel[]);
+
+export const selectedChannelIndexAtom = atom(ChannelSelection.All);
