+++ src/components/video/VideoGrid.tsx
@@ -3,22 +3,19 @@ import Grid from '@material-ui/core/Grid';
 import Box from '@material-ui/core/Box';
 import Skeleton from '@material-ui/lab/Skeleton';
 import { Video } from '../../models/Video';
-import Media from './Media';
+import VideoRenderer from './VideoRenderer';
 import { debug } from '../../helpers/debug';
 import { styles } from './VideoGrid.styles';
 
 interface VideoGridProps {
   loading?: boolean;
-  selectedChannelIndex: number;
   videos: Video[];
   maxPerChannel?: number;
   maxSkeletons?: number;
-  onVideoClick: Function;
-  onVideoWatchLaterClick: Function;
 }
 
 export default function VideoGrid(props: VideoGridProps) {
-  const { selectedChannelIndex, videos, loading = false, maxPerChannel = 9, maxSkeletons = 9, onVideoClick, onVideoWatchLaterClick } = props;
+  const { videos, loading = false, maxPerChannel = 9, maxSkeletons = 9 } = props;
   const [preventLongPress, setPreventLongPress] = React.useState(false);
 
   const handleMouseEvent = (event: any) => {
@@ -43,15 +40,10 @@ export default function VideoGrid(props: VideoGridProps) {
 
   return (
     <Grid container style={styles.grid} onMouseDown={(event: any) => handleMouseEvent(event)} onClickCapture={(event: any) => handleMouseEvent(event)}>
-      {(loading ? Array.from(new Array(Math.min(maxPerChannel, maxSkeletons))) : videos.slice(0, maxPerChannel)).map((item, index) => (
+      {(loading ? Array.from(new Array(Math.min(maxPerChannel, maxSkeletons))) : videos.slice(0, maxPerChannel)).map((video, index) => (
         <Box key={index} width={210} marginRight={0.5} marginBottom={3} draggable="false">
-          {item ? (
-            <Media
-              item={item}
-              selectedChannelIndex={selectedChannelIndex}
-              onClick={onVideoClick}
-              onWatchLaterClick={onVideoWatchLaterClick}
-            ></Media>
+          {video ? (
+            <VideoRenderer video={video} ></VideoRenderer>
           ) : (
             <React.Fragment>
               <Skeleton variant="rect" width={210} height={118} />
